+++ include/board.hpp
@@ -18,30 +18,30 @@ namespace PijersiEngine
         Board(Board &board);
         // ~Board();
 
-        void playManual(std::vector<uint32_t> move);
-        void playManual(uint32_t move);
+        void playManual(std::vector<uint64_t> move);
+        void playManual(uint64_t move);
         void playManual(std::string move);
 
         // Depth limited search
 
-        uint32_t searchDepth(int recursionDepth, bool random, uint32_t prinvipalVariation = NULL_MOVE, uint64_t searchTimeMilliseconds = UINT64_MAX, bool iterative = true);
-        uint32_t playDepth(int recursionDepth, bool random = true, uint32_t prinvipalVariation = NULL_MOVE, uint64_t searchTimeMilliseconds = UINT64_MAX, bool iterative = true);
+        uint64_t searchDepth(int recursionDepth, bool random, uint64_t prinvipalVariation = NULL_MOVE, uint64_t searchTimeMilliseconds = UINT64_MAX, bool iterative = true);
+        uint64_t playDepth(int recursionDepth, bool random = true, uint64_t prinvipalVariation = NULL_MOVE, uint64_t searchTimeMilliseconds = UINT64_MAX, bool iterative = true);
 
         // Time limited search
         
-        uint32_t searchTime(bool random, uint64_t searchTimeMilliseconds = UINT64_MAX);
-        uint32_t playTime(bool random = true, uint64_t searchTimeMilliseconds = UINT64_MAX);
+        uint64_t searchTime(bool random, uint64_t searchTimeMilliseconds = UINT64_MAX);
+        uint64_t playTime(bool random = true, uint64_t searchTimeMilliseconds = UINT64_MAX);
         
         // Random search
 
-        uint32_t searchRandom();
-        uint32_t playRandom();
+        uint64_t searchRandom();
+        uint64_t playRandom();
 
-        // uint32_t playMCTS(int seconds = 10, int simulationsPerRollout = 3);
-        // uint32_t ponderMCTS(int seconds, int simulationsPerRollout);
+        // uint64_t playMCTS(int seconds = 10, int simulationsPerRollout = 3);
+        // uint64_t ponderMCTS(int seconds, int simulationsPerRollout);
         
         std::string advice(int recursionDepth, bool random);
-        bool isMoveLegal(uint32_t move);
+        bool isMoveLegal(uint64_t move);
         int64_t evaluate();
         
         void setStringState(std::string stateString);
@@ -64,14 +64,14 @@ namespace PijersiEngine
         uint8_t cells[45];
 
     private:
-        uint32_t searchBook();
+        uint64_t searchBook();
 
-        uint32_t countPieces();
+        uint64_t countPieces();
         void endTurn();
 
-        uint32_t lastPieceCount = 0;
-        uint32_t halfMoveCounter = 0;
-        uint32_t moveCounter = 1;
+        uint64_t lastPieceCount = 0;
+        uint64_t halfMoveCounter = 0;
+        uint64_t moveCounter = 1;
 
         void addPiece(uint8_t piece, int i, int j);
     };
